0-linear.c

+++ 0x1E-search_algorithms/0-linear.c
@@ -10,11 +10,11 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-size_t i = 0;
+int i = 0;
 
-for (i = 0; i < size; i++)
+for (i = 0; i < (int)size; i++)
 {
-printf("Value checked array[%ld] = %d\n", i, array[i]);
+printf("Value checked array[%u] = %d\n", i, array[i]);
 if (array[i] == value)
 {
 return (i);
